Update LegendItem.py
There was an error where a legend had a property `legend.offset`, which overwrites the correct method `legend.offset()`, and is itself never actually updated or used, so is incorrect. When you call `legend.offset()` you get an error, since you then call the property `legend.offset`, which is not a function. This is fixed by removing `self.offset = offset`.